Refactor Packages to UI and Data Layer
