+++ app/src/main/java/com/survivalcoding/todolist/ui/main/adapter/TodoDiffItemCallback.kt
@@ -1,7 +1,7 @@
-package com.survivalcoding.todolist.presentation.main.adapter
+package com.survivalcoding.todolist.ui.main.adapter
 
 import androidx.recyclerview.widget.DiffUtil
-import com.survivalcoding.todolist.model.Todo
+import com.survivalcoding.todolist.data.model.Todo
 
 object TodoDiffItemCallback : DiffUtil.ItemCallback<Todo>() {
     override fun areItemsTheSame(oldItem: Todo, newItem: Todo): Boolean {
