+++ app/src/main/java/com/survivalcoding/todolist/ui/main/adapter/TodoListAdapter.kt
@@ -1,10 +1,10 @@
-package com.survivalcoding.todolist.presentation.main.adapter
+package com.survivalcoding.todolist.ui.main.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import com.survivalcoding.todolist.databinding.ListItemTodoBinding
-import com.survivalcoding.todolist.model.Todo
+import com.survivalcoding.todolist.data.model.Todo
 
 class TodoListAdapter(private val onClick: TodoListAdapter.(Long) -> Unit) :
     ListAdapter<Todo, TodoViewHolder>(TodoDiffItemCallback) {
