+++ app/src/main/res/layout/fragment_todo.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.MainActivity"
+    tools:context=".ui.MainActivity"
     android:padding="16dp">
 
     <CheckBox
